AI/requirements.txt
flask-cors
Flask[async]
pymupdf
google-generativeai
langchain
faiss-cpu
langchain
langchain-google-genai
gunicorn
Pillow
torch
typing-extensions
transformers
numpy
opencv-python